requirements.txt
altair==6.0.0
altex==0.2.0
annotated-types==0.7.0
anyio==4.12.0
asn1crypto==1.5.1
asttokens==3.0.1
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
boto3==1.42.5
botocore==1.42.5
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.1
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
distro==1.9.0
entrypoints==0.4
executing==2.2.1
faker==38.2.0
favicon==0.7.0
filelock==3.20.0
fonttools==4.61.0
gitdb==4.0.12
gitpython==3.1.45
greenlet==3.3.0
h11==0.16.0
htbuilder==0.9.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ipykernel==7.1.0
ipython==9.8.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-client==8.6.3
jupyter-core==5.9.1
kiwisolver==1.4.9
lxml==6.0.2
markdown==3.10
markdownlit==0.0.7
markupsafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
narwhals==2.13.0
nest-asyncio==1.6.0
numpy==2.3.5
openai==2.9.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pathlib==1.0.1
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.1
playwright==1.57.0
plotly==6.5.0
prometheus-client==0.23.1
prompt-toolkit==3.0.52
protobuf==6.31.1
psutil==7.1.3
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==22.0.0
pycparser==2.23
pydantic==2.12.5
pydantic-core==2.41.5
pydeck==0.9.1
pyee==13.0.0
pygments==2.19.2
pyjwt==2.10.1
pymdown-extensions==10.18
pyopenssl==25.3.0
pyparsing==3.2.5
pysocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.3
pyzmq==27.1.0
referencing==0.37.0
reportlab==4.4.6
requests==2.32.5
rpds-py==0.30.0
s3transfer==0.16.0
selenium==4.39.0
setuptools==80.9.0
shapely==2.1.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowflake-connector-python==4.1.1
snowflake-snowpark-python==1.43.0
sortedcontainers==2.4.0
soupsieve==2.8
st-annotated-text==4.0.2
st-theme==1.2.3
stack-data==0.6.3
streamlit==1.52.1
streamlit-autorefresh==1.0.1
streamlit-avatar==0.1.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.7.8
streamlit-faker==0.0.4
streamlit-image-coordinates==0.4.0
streamlit-keyup==0.3.0
streamlit-notify==0.3.1
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
tenacity==9.1.2
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.32.0
trio-websocket==0.12.2
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.6.1
validators==0.35.0
watchdog==6.0.0
wcwidth==0.2.14
websocket-client==1.9.0
wheel==0.45.1
wsproto==1.3.2
